// projects/codx-ac/src/lib/models/ItemBatchs.model.ts
export class ItemBatchs{
    recID: any;
    itemID: any;
    batchNo: any;
    batchName: any;
    idiM0: any;
    idiM1: any;
    idiM2: any;
    idiM3: any;
    idiM4: any;
    idiM5: any;
    idiM7: any;
    idiM8: any;
    idiM9: any;
    manufacturer: any;
    manufaturedDate: any;
    expiredDate: any;
    bestBeforeDate: any;
    statusCodeID: any;
    available: any;
    status: any;
    parentBatchNo: any;
    vendorBatchNo: any;
    vendorBatchDate: any;
    vendorExpiredDate: any;
    useVendorExpiredDate: any;
    cwUM: any;
    cwConversion: any;
    cwQty: any;
    qty: any;
    qtyA: any;
    qtyB: any;
    qtyC: any;
    nWeight: any;
    gWeight: any;
    tWeight: any;
    tCTStd: any;
    tCTAct: any;
    siteID: any;
    prodStageID: any;
    wcID: any;
    machineID: any;
    shiftID: any;
    terminalID: any;
    qCEmployee: any;
    qCNote: any;
    potency: any;
    density: any;
    coneType: any;
    receiptNo: any;
    receiptDate: any;
    issueNo: any;
    issueDate: any;
    customerID: any;
    orderNo: any;
    printedCode: any;
    splitedCoils: any;
    lastNo: any;
    note: any;
    stop: any;
    autoCreated: any;
    owner: any;
    buid: any;
    createdOn: any;
    createdBy: any;
    modifiedOn: any;
    modifiedBy: any;
}